refactor: stick to provider component
