+++ src/components/AuthGuardProvider.tsx
@@ -1,13 +1,23 @@
-import React from "react";
+import React, { useEffect } from "react";
+import { useLocation, useNavigate } from "react-router-dom";
 
-import { useAuthGuard } from "@/hooks/useAuthGuard";
+import { useAuthQuery } from "@/apis/oauth";
 
 interface AuthGuardProviderProps extends React.PropsWithChildren {
-  redirectTo?: string;
+  options?: { redirectTo?: string };
 }
 
-export default function AuthGuardProvider({ children, redirectTo }: AuthGuardProviderProps) {
-  useAuthGuard({ redirectTo });
+export default function AuthGuardProvider({ children, options }: AuthGuardProviderProps) {
+  const navigate = useNavigate();
+  const { data: isAuthenticated } = useAuthQuery();
+  const redirectTo = options?.redirectTo || "/discordlogin";
+  const location = useLocation();
+
+  useEffect(() => {
+    if (!isAuthenticated && location.pathname !== redirectTo) {
+      navigate(redirectTo, { replace: true });
+    }
+  }, [isAuthenticated, navigate, redirectTo, location.pathname]);
 
   return <>{children}</>;
 }
